app/requirements.txt
streamlit==1.31.0
pandas==2.2.0
pandas-gbq==0.22.0
google-cloud-bigquery==3.14.1
pyarrow==15.0.0
